Add support of report files (with log always) and add button to download archive with these files.

+++ bridge/tools/utils.py
@@ -14,9 +14,8 @@ def clear_job_files():
     from jobs.models import File, JOBFILE_DIR
     files_in_the_system = []
     for f in File.objects.all():
-        if len(f.etvfiles_set.all()) == 0 \
-                and len(f.reportcomponent_set.all()) == 0 \
-                and len(f.filesystem_set.all()) == 0:
+        if len(f.etvfiles_set.all()) == 0 and len(f.reportcomponent_set.all()) == 0 \
+                and len(f.filesystem_set.all()) == 0 and len(f.reportfiles_set.all()) == 0:
             f.delete()
         else:
             file_path = os.path.abspath(os.path.join(MEDIA_ROOT, f.file.name))
